[PIE-0]: Add limit to the orchestration log

+++ 860-orchestration-visualization/src/main/java/io/harness/event/OrchestrationEndGraphHandler.java
@@ -50,15 +50,14 @@ public class OrchestrationEndGraphHandler implements AsyncInformObserver, Orches
   public void onEnd(Ambiance ambiance) {
     try (AutoLogContext autoLogContext = AmbianceUtils.autoLogContext(ambiance)) {
       PlanExecution planExecution = planExecutionService.get(ambiance.getPlanExecutionId());
-      try {
-        // One last time try to update the graph to process any unprocessed logs
-        graphGenerationService.updateGraphWithWaitLock(planExecution.getUuid());
-        // We are not deleting logs pro-actively if exception occurred, they will be helpful in debugging.
-        orchestrationEventLogRepository.deleteLogsForGivenPlanExecutionId(ambiance.getPlanExecutionId());
-      } catch (Exception ex) {
+      // One last time try to update the graph to process any unprocessed logs
+      if (!graphGenerationService.updateGraphWithWaitLock(planExecution.getUuid())) {
         log.error(
             "[GRAPH_ERROR] Exception occurred while updating graph through logs. Regenerating the graph from nodeExecutions");
         graphGenerationService.buildOrchestrationGraph(ambiance.getPlanExecutionId());
+      } else {
+        // We are not deleting logs pro-actively if exception occurred, they will be helpful in debugging.
+        orchestrationEventLogRepository.deleteLogsForGivenPlanExecutionId(ambiance.getPlanExecutionId());
       }
 
       // Todo: Check if this is required
